requirements.txt
fastapi
uvicorn[standard]
pydantic
starlette
pyjwt
aiosqlite
pipmaster
lollms_client
openai